requirements.txt
streamlit==1.39.0
sentence-transformers==3.1.1
pypdf2==3.0.1
pytesseract==0.3.13
pillow==10.4.0
opensearch-py==2.7.1
torch==2.4.1
numpy==2.1.2
requests==2.32.3
ollama==0.3.3
openpyxl==3.1.5
python-pptx==1.0.2
